extracted featured image url logic
